chore: use Double for price and refine catalog list UI/state

+++ composeApp/src/commonMain/kotlin/org/example/kmp/feature/catalog/api/dto/ProductDto.kt
@@ -8,7 +8,7 @@ data class ProductDto(
     @SerialName("id") val id: Int = 0,
     @SerialName("title") val title: String = "",
     @SerialName("description") val description: String = "",
-    @SerialName("price") val price: Int = 0,
+    @SerialName("price") val price: Double = 0.0,
     @SerialName("thumbnail") val thumbnailUrl: String = "",
     @SerialName("images") val images: List<String> = emptyList(),
 )
